% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOPdim2.R
\name{data_gen2D}
\alias{data_gen2D}
\title{data_gen2D}
\usage{
data_gen2D(n, chpts = NULL, means1 = 0, means2 = 0, noise = 1)
}
\arguments{
\item{n}{number of data point.}

\item{chpts}{a vector of increasing changepoint indices. Last index is always less than 'n'.
By default, 'chpts = NULL' (the data without changepoints).}

\item{means1}{vector of successive means1 for data1, by default 'means1 = 0'.}

\item{means2}{vector of successive means2 for data2, by default 'means2 = 0'.}

\item{noise}{standard deviation of an additional normal noise, by default 'noise = 1'.}
}
\value{
matrix of data of dimension 2 x n with a given values of means by the segmentation.
}
\description{
Generation of data of dimension 2 with a given values of means and changepoints
}
\examples{
Data <- data_gen2D(n = 10, chpts = NULL, means1 = 0, means2 = 0, noise = 1)
}
