% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOPdim2.R
\name{plotFPOP2D}
\alias{plotFPOP2D}
\title{plotFPOP2D}
\usage{
plotFPOP2D(data, chpts, means1, means2)
}
\arguments{
\item{data}{matrix of data of dimension 2 x data length.}

\item{chpts}{a vector of increasing changepoint indices (last element is always n!).}

\item{means1}{vector of successive means1 for data1.}

\item{means2}{vector of successive means2 for data2.}
}
\value{
Plot of data with a given values of means and changepoints.
}
\description{
Plot of data with a  values of means and changepoints.
}
\examples{
Data <- data_gen2D(n = 10, chpts = c(10), means1 = c(0), means2 =c(1), noise = 1)
plotFPOP2D(data = Data, chpts= c(10), means1 = c(0), means2 = c(1))
}
