% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOPdim2.R
\name{data_gen2D}
\alias{data_gen2D}
\title{data_gen2D}
\usage{
data_gen2D(n, chpts, means1, means2, noise = 1)
}
\arguments{
\item{n}{number of data point.}

\item{chpts}{a vector of increasing changepoint indices (last element is always n!).}

\item{means1}{vector of successive means1 for data1.}

\item{means2}{vector of successive means2 for data2.}

\item{noise}{standard deviation of an additional normal noise.}
}
\value{
matrix of data of dimension 2 x n with a given values of means by the segmentation.
}
\description{
Generation of data of dimension 2 with a given values of means and changepoints
}
\examples{
data_gen2D(n = 10, chpts = c(10), means1 = c(0), means2 =c(1), noise = 1)
}
